增加下位机代码，基本功能联调通过

+++ pc-sdk/csharp/NexLinkPacket.cs
@@ -7,7 +7,7 @@ namespace NexLink
     {
         public byte magic;
         public byte type;
-        public byte cmd;
+        public ushort cmd;
         public ushort seq;
         public ushort length;
